Move codegen api client into its own directory
