+++ packages/api/src/client/types/app/bsky/badge.ts
