+++ packages/api/src/client/types/app/bsky/badgeAccept.ts
