+++ packages/api/src/client/types/app/bsky/badgeOffer.ts
