+++ packages/api/src/client/types/app/bsky/postNotificationsSeen.ts
