+++ packages/api/src/client/types/com/atproto/getAccountsConfig.ts
