+++ packages/api/src/client/types/com/atproto/repoGetRecord.ts
