+++ packages/api/src/client/types/com/atproto/syncGetRoot.ts
